% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_countpoint_data.R
\name{filter_countpoint_data}
\alias{filter_countpoint_data}
\title{Filters countpoint_aadf, countpoint_raw, countpoint_aadf_bydirection or countpoint_raw_bydirection}
\usage{
filter_countpoint_data(
  df1,
  df2,
  df3,
  plot_type,
  traffic_class,
  year_from,
  year_to,
  estimation_method
)
}
\arguments{
\item{df1}{dataframe any direction}

\item{df2}{dataframe by direction}

\item{df3}{dataframe by day of week}

\item{plot_type}{character}

\item{traffic_class}{character}

\item{year_from}{integer}

\item{year_to}{integer}

\item{estimation_method}{character}
}
\value{
Returns dataframe countpoint_filtered
}
\description{
Depends on countpoint_plot_type and filters on traffic_class, countpoint_year_from
and countpoint_estimation_method
}
