% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_traffic_data.R
\name{filter_traffic_data}
\alias{filter_traffic_data}
\title{Generates traffic_filtered from la_traffic}
\usage{
filter_traffic_data(la_traffic, tv_area, tv_num_areas)
}
\arguments{
\item{la_traffic}{dataframe}

\item{tv_area}{character}

\item{tv_num_areas}{integer}
}
\value{
Returns dataframe traffic_filtered. This will be in wide format if it contains more
than one area (single % change line to be plotted for each area). For a single area it will be
in long format (multiple lines to be plotted for the same area).
}
\description{
Depends on the area(s) selected by the user in tv_area and the
the number of areas selected. If more than 1 group by area and
create the % diff columns for each mode; if one area then la_traffic
is pivoted long to enable the display of multiple lines for each vehicle
type and mode.
}
