% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gbtraffic_info_link}
\alias{gbtraffic_info_link}
\title{Generate information icon}
\usage{
gbtraffic_info_link(text, id, size = "standard")
}
\arguments{
\item{text}{string, text to be used for the label}

\item{id}{string, id to identify the actionLink when clicked}

\item{size}{string, size of icon to be used for the actionLink}
}
\value{
Returns a list of two items to be used as an input label: the passed text and an icon-based actionLink
}
\description{
Generate information icon
}
\keyword{internal}
