% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gbtraffic_info_modal}
\alias{gbtraffic_info_modal}
\title{Show a modal}
\usage{
gbtraffic_info_modal(title = NULL, file)
}
\arguments{
\item{title}{string, text to be used for the model title}

\item{file}{string, file name of the html file with modal content}
}
\description{
Show a modal
}
\keyword{internal}
