% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gbtraffic_line_plot}
\alias{gbtraffic_line_plot}
\title{Generate line plot}
\usage{
gbtraffic_line_plot(
  df,
  x,
  y,
  group,
  title = NULL,
  subtitle = NULL,
  title_face = NULL,
  caption = NULL,
  title_size = NULL,
  subtitle_size = NULL,
  legend_title = NULL,
  legend_title_size = NULL,
  legend_title_face = NULL,
  legend = TRUE,
  legend_nrow = 1,
  highlight = NULL,
  highlight_size = 1,
  x_tick_label_size = NULL,
  y_tick_label_size = NULL,
  x_label = NULL,
  y_label = NULL,
  colour_palette = "Set1"
)
}
\arguments{
\item{df}{data frame}

\item{x}{character, name of column for x axis. Can also be a function to be applied to that column, e.g "factor(year)".}

\item{y}{character, name of column for x axis. Can also be a function to be applied to that column.}

\item{group}{character, name of column to partition by. Can also be a function to be applied to that column, e.g. "factor(traffic_area_id)".}

\item{title}{character.}

\item{title_face}{character, font face.}

\item{title_size}{integer, font size to use for plot title.}

\item{subtitle_size}{integer, font size to use for plot subtitle.}

\item{legend_title}{character.}

\item{legend_title_size}{integer font size to use for legend title.}

\item{legend_title_face}{character, font size to use for legend title.}

\item{legend}{Boolean, whether a legend should be shown. Default is TRUE.}

\item{legend_nrow}{integer, number of rows to use for the legend. Default is 1.}

\item{highlight}{character, single value from the group column where specific line thickness is to be set.}

\item{highlight_size}{integer, weight of a line to be highlighted.}

\item{x_tick_label_size}{integer, font size of the x tick labels.}

\item{y_tick_label_size}{integer, font size of the y tick labels.}

\item{x_label}{character, label for the x axis.}

\item{y_label}{character, label for the x axis.}

\item{colour_palette}{character, name of a Color Brewer palette from brewer.pal.info.}
}
\description{
Generate line plot
}
\keyword{internal}
