% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gbtraffic_rtable}
\alias{gbtraffic_rtable}
\title{Generate a reactable table}
\usage{
gbtraffic_rtable(
  df,
  columns,
  columnGroups = NULL,
  groupBy = NULL,
  bordered = TRUE,
  striped = TRUE,
  compact = TRUE,
  showSortable = FALSE,
  defaultPageSize = 5,
  showPageSizeOptions = TRUE,
  pageSizeOptions = c(5, 10, 20, 40),
  defaultColDef = colDef(minWidth = 50)
)
}
\arguments{
\item{df}{dataframe}

\item{columns}{list of columns}

\item{columnGroups}{list of column groups}

\item{groupBy}{character column to group by}

\item{bordered}{boolean}

\item{striped}{boolean}

\item{compact}{boolean}

\item{showSortable}{boolean}

\item{defaultPageSize}{integer}

\item{showPageSizeOptions}{boolean}

\item{pageSizeOptions}{vector of integers}

\item{defaultColDef}{colDef object}
}
\description{
Generate a reactable table
}
\keyword{internal}
