% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countpoint_data.R
\name{get_countpoint_data}
\alias{get_countpoint_data}
\title{Gets the data for the selected countpoint}
\usage{
get_countpoint_data(pool, count_id, count_type)
}
\arguments{
\item{pool}{pool object}

\item{count_id}{integer or character}

\item{count_type}{text}
}
\description{
Gets the data for the selected countpoint
}
