% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countpoint_info.R
\name{get_countpoint_info}
\alias{get_countpoint_info}
\title{Gets information about the selected countpoint}
\usage{
get_countpoint_info(df, count_id, count_type)
}
\arguments{
\item{df}{dataframe of countpoints}

\item{count_id}{integer or character}

\item{count_type}{text}
}
\description{
Gets information about the selected countpoint
}
