% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_countpoint_plot.R
\name{output_countpoint_plot}
\alias{output_countpoint_plot}
\title{Generates the bar chart for output$countpoint_plot using countpoint_aadf_filtered}
\usage{
output_countpoint_plot(
  df,
  plot_type,
  estimation_method,
  traffic_class,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{df}{dataframe of filtered countpoint data}

\item{plot_type}{character}

\item{estimation_method}{character}

\item{traffic_class}{character}

\item{title}{character}

\item{subtitle}{character}

\item{caption}{character}
}
\value{
Returns ggplot object
}
\description{
Validates that there are rows to display; validates that there are rows for
the currently selected class and the type of bar chart generated depends on
aadf_plot_type.
}
