% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_roadlength_plot.R
\name{output_roadlength_plot}
\alias{output_roadlength_plot}
\title{Generates road length plots for the traffic volume tab.}
\usage{
output_roadlength_plot(traffic_filtered, road_type)
}
\arguments{
\item{traffic_filtered}{dataframe}

\item{road_type}{string}
}
\value{
Returns ggplot object.
}
\description{
Generates road length plots for the traffic volume tab.
}
