% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_traffic_plot.R
\name{output_traffic_plot}
\alias{output_traffic_plot}
\title{Generates plots for the traffic volume tab.}
\usage{
output_traffic_plot(
  traffic_filtered,
  tv_num_areas,
  tv_type,
  tv_type_cn,
  tv_area
)
}
\arguments{
\item{traffic_filtered}{dataframe}

\item{tv_num_areas}{integer}

\item{tv_type}{character. Traffic road type and mode}

\item{tv_type_cn}{character. Common name of traffic road type and mode}

\item{tv_area}{character}
}
\value{
Returns ggplot object.
}
\description{
Plot type depends on the number of areas selected by the user (tv_num_areas).
}
