% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_traffic_table.R
\name{output_traffic_table}
\alias{output_traffic_table}
\title{Generates tables for the traffic volume tab from traffic_filtered}
\usage{
output_traffic_table(traffic_filtered, tv_num_areas)
}
\arguments{
\item{traffic_filtered}{dataframe}

\item{tv_num_areas}{integer}
}
\value{
Returns a reactable object
}
\description{
Table generated depends on the number of areas selected by user.
}
