% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input object}

\item{output}{output object}

\item{session}{session object}
}
\description{
Server function
}
